requirements.txt
beautifulsoup4==4.14.3
requests==2.33.1
python-dotenv==1.2.2
